# public/swagger.json
{
	"components": {
...
		"schemas": {
			"ClientModel": {
				"properties": {
					"clientcode": {
						"type": "string"
					},
					"street": {
						"type": "string"
					},
					"city": {
						"type": "string"
					},
					"pob": {
						"type": "string"
					},
					"zipcode": {
						"type": "string"
					},
					"pobzipcode": {
						"type": "string"
					},
					"longitude": {
						"type": "string"
					},
					"latitude": {
						"type": "string"
					},
					"country": {
						"type": "string"
					},
					"clientname": {
						"type": "string"
					}
				},
				"required": [
					"clientcode",
					"street",
					"city",
					"pob",
					"zipcode",
					"pobzipcode",
					"longitude",
					"latitude",
					"country",
					"clientname"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ShipmentModel": {
				"properties": {
					"shipmentid": {
						"type": "number",
						"format": "double"
					},
					"sequencenumber": {
						"type": "number",
						"format": "double"
					},
					"date": {
						"type": "string",
						"format": "date-time"
					},
					"modality": {
						"type": "string"
					},
					"distance": {
						"type": "number",
						"format": "double"
					},
					"intervaldistance": {
						"type": "string"
					},
					"emissionprocessing": {
						"type": "number",
						"format": "double"
					},
					"fefrom": {
						"type": "string"
					},
					"feto": {
						"type": "string"
					},
					"routeid": {
						"type": "string"
					},
					"fromclient": {
						"type": "string"
					},
					"toclient": {
						"type": "string"
					},
					"containertype": {
						"type": "string"
					},
					"intervalweight": {
						"type": "string"
					},
					"transporttype": {
						"type": "string"
					},
					"shipmentweight": {
						"type": "number",
						"format": "double"
					},
					"teu": {
						"type": "number",
						"format": "double"
					},
					"co2pershipment": {
						"type": "number",
						"format": "double"
					},
					"Container": {
						"$ref": "#/components/schemas/ContainerModel"
					},
					"CO2": {
						"$ref": "#/components/schemas/CO2Model"
					},
					"NOX": {
						"$ref": "#/components/schemas/NOXModel"
					}
				},
				"required": [
					"shipmentid",
					"sequencenumber",
					"date",
					"modality",
					"distance",
					"intervaldistance",
					"emissionprocessing",
					"fefrom",
					"feto",
					"routeid",
					"fromclient",
					"toclient",
					"containertype",
					"intervalweight",
					"transporttype",
					"shipmentweight",
					"teu",
					"co2pershipment",
					"Container",
					"CO2",
					"NOX"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ContainerModel": {
				"properties": {
					"containertype": {
						"type": "string"
					},
					"lengthcontainer": {
						"type": "number",
						"format": "double"
					},
					"widthcontainer": {
						"type": "number",
						"format": "double"
					},
					"heightcontainer": {
						"type": "number",
						"format": "double"
					},
					"Shipment": {
						"items": {
							"$ref": "#/components/schemas/ShipmentModel"
						},
						"type": "array"
					}
				},
				"required": [
					"containertype",
					"lengthcontainer",
					"widthcontainer",
					"heightcontainer",
					"Shipment"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CO2Model": {
				"properties": {
					"emissionsperkm": {
						"type": "number",
						"format": "double"
					},
					"transporttype": {
						"type": "string"
					},
					"fe": {
						"type": "string"
					},
					"intervalweight": {
						"type": "string"
					},
					"teu": {
						"type": "number",
						"format": "double"
					},
					"Shipment": {
						"items": {
							"$ref": "#/components/schemas/ShipmentModel"
						},
						"type": "array"
					}
				},
				"required": [
					"emissionsperkm",
					"transporttype",
					"fe",
					"intervalweight",
					"teu",
					"Shipment"
				],
				"type": "object",
				"additionalProperties": false
			},
			"NOXModel": {
				"properties": {
					"emissionsperkm": {
						"type": "number",
						"format": "double"
					},
					"transporttype": {
						"type": "string"
					},
					"intervalweight": {
						"type": "string"
					},
					"teu": {
						"type": "number",
						"format": "double"
					},
					"Shipment": {
						"items": {
							"$ref": "#/components/schemas/ShipmentModel"
						},
						"type": "array"
					}
				},
				"required": [
					"emissionsperkm",
					"transporttype",
					"intervalweight",
					"teu",
					"Shipment"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RoutesModel": {
				"properties": {
					"routeid": {
						"type": "string"
					},
					"modality": {
						"type": "string"
					},
					"fueltype": {
						"type": "string"
					},
					"TransportType": {
						"type": "string"
					}
				},
				"required": [
					"routeid",
					"modality",
					"fueltype",
					"TransportType"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "api",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/client/{clientCode?}": {
			"get": {
				"operationId": "GetClients",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ClientModel"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Client"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "offset",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "path",
						"name": "clientCode",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/client": {
			"post": {
				"operationId": "CreateClient",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ClientModel"
								}
							}
						}
					}
				},
				"tags": [
					"Client"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ClientModel"
							}
						}
					}
				}
			},
			"put": {
				"operationId": "UpdateClient",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ClientModel"
								}
							}
						}
					}
				},
				"tags": [
					"Client"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ClientModel"
							}
						}
					}
				}
			}
		},
		"/client/{clientCode}": {
			"delete": {
				"operationId": "DeleteClient",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ClientModel"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Client"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "clientCode",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/co2": {
			"get": {
				"operationId": "GetCo2s",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/CO2Model"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"CO2"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "offset",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateCo2",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CO2Model"
								}
							}
						}
					}
				},
				"tags": [
					"CO2"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CO2Model"
							}
						}
					}
				}
			},
			"put": {
				"operationId": "UpdateCo2",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CO2Model"
								}
							}
						}
					}
				},
				"tags": [
					"CO2"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CO2Model"
							}
						}
					}
				}
			}
		},
		"/co2/query": {
			"get": {
				"operationId": "GetCo2",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/CO2Model"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"CO2"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "fe",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "intervalweight",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "teu",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "transporttype",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteCo2",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/CO2Model"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"CO2"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "fe",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "intervalweight",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "teu",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "transporttype",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/container/{containerType?}": {
			"get": {
				"operationId": "GetContainers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ContainerModel"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Container"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "offset",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "path",
						"name": "containerType",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/container": {
			"post": {
				"operationId": "CreateContainer",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ContainerModel"
								}
							}
						}
					}
				},
				"tags": [
					"Container"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ContainerModel"
							}
						}
					}
				}
			},
			"put": {
				"operationId": "UpdateContainer",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ContainerModel"
								}
							}
						}
					}
				},
				"tags": [
					"Container"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ContainerModel"
							}
						}
					}
				}
			}
		},
		"/container/{containerType}": {
			"delete": {
				"operationId": "DeleteContainer",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ContainerModel"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Container"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "containerType",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/nox": {
			"get": {
				"operationId": "GetNoxs",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/NOXModel"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Nox"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "offset",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateNox",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NOXModel"
								}
							}
						}
					}
				},
				"tags": [
					"Nox"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/NOXModel"
							}
						}
					}
				}
			},
			"put": {
				"operationId": "UpdateNox",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NOXModel"
								}
							}
						}
					}
				},
				"tags": [
					"Nox"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/NOXModel"
							}
						}
					}
				}
			}
		},
		"/nox/query": {
			"get": {
				"operationId": "GetNox",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/NOXModel"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Nox"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "intervalweight",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "teu",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "transporttype",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteNox",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/NOXModel"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Nox"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "intervalweight",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "teu",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "transporttype",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/Route/{routeId?}": {
			"get": {
				"operationId": "GetRoutes",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/RoutesModel"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Route"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "offset",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "path",
						"name": "routeId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/Route": {
			"post": {
				"operationId": "CreateRoute",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RoutesModel"
								}
							}
						}
					}
				},
				"tags": [
					"Route"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RoutesModel"
							}
						}
					}
				}
			},
			"put": {
				"operationId": "UpdateRoute",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RoutesModel"
								}
							}
						}
					}
				},
				"tags": [
					"Route"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RoutesModel"
							}
						}
					}
				}
			}
		},
		"/Route/{routeId}": {
			"delete": {
				"operationId": "DeleteRoute",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/RoutesModel"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Route"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "routeId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}